@@ -1,14 +1,11 @@
 module github.com/mikeknep/jeru
 
-go 1.15
+go 1.16
 
 require (
-	github.com/AlecAivazis/survey/v2 v2.2.12 // indirect
-	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/AlecAivazis/survey/v2 v2.2.12
 	github.com/manifoldco/promptui v0.8.0
-	github.com/mitchellh/go-homedir v1.1.0
-	github.com/pterm/pterm v0.12.13 // indirect
+	github.com/pterm/pterm v0.12.13
 	github.com/spf13/cobra v1.1.1
-	github.com/spf13/viper v1.7.1
 	github.com/stretchr/testify v1.7.0
 )
